+++ src/i18next.api.js
@@ -2,6 +2,7 @@
 i18n.init = init;
 i18n.setLng = setLng;
 i18n.preload = preload;
+i18n.addResourceBundle = addResourceBundle;
 i18n.loadNamespace = loadNamespace;
 i18n.loadNamespaces = loadNamespaces;
 i18n.setDefaultNamespace = setDefaultNamespace;